Update core.make
It seems like many students gets stuck due to errors when performing "make xsch". Giving them easier access to the output of Xschem in the form of a log file will make it easier to figure out what is causing the errors and will keep the terminal clean. Also, it will teach them to troubleshoot using log files and to become more independent.